pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "metochina"
version = "1.2.0"
description = "OSINT Image Metadata Analysis Tool — extract, analyze, and report hidden metadata in images"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Belchite"},
]
keywords = ["osint", "exif", "metadata", "gps", "forensics", "privacy", "image-analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]
dependencies = [
    "Pillow>=10.0",
    "click>=8.1",
]

[project.scripts]
metochina = "metochina.cli:cli"

[project.urls]
Homepage = "https://github.com/metochina/metochina"
Repository = "https://github.com/metochina/metochina"
Issues = "https://github.com/metochina/metochina/issues"

[tool.setuptools.packages.find]
include = ["metochina*"]
